+++ src/dsp_tools/commands/xmlupload/resource_multimedia.py
@@ -9,8 +9,8 @@ from loguru import logger
 from dsp_tools.commands.xmlupload.models.deserialise.deserialise_value import XMLBitstream
 from dsp_tools.commands.xmlupload.models.deserialise.xmlresource import BitstreamInfo
 from dsp_tools.commands.xmlupload.models.deserialise.xmlresource import XMLResource
+from dsp_tools.commands.xmlupload.models.ingest import IngestClient
 from dsp_tools.commands.xmlupload.models.permission import Permissions
-from dsp_tools.commands.xmlupload.models.sipi import IngestClient
 from dsp_tools.models.exceptions import PermanentConnectionError
